chore: rename NavStack to Navigation; uncomment stack navigation
